# Backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PMS Admin API",
    "description": "Auto-generated API documentation for PMS Admin Portal",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
  "paths": {
...
    "/api/clients/clients": {
      "post": {
        "tags": [
          "Clients"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "company_name": {
                    "example": "any"
                  },
                  "company_type": {
                    "example": "any"
                  },
                  "company_size": {
                    "example": "any"
                  },
                  "company_industry": {
                    "example": "any"
                  },
                  "tax_info": {
                    "example": "any"
                  },
                  "address_line1": {
                    "example": "any"
                  },
                  "address_line2": {
                    "example": "any"
                  },
                  "country": {
                    "example": "any"
                  },
                  "state": {
                    "example": "any"
                  },
                  "city": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Clients"
        ],
        "description": "",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/clients/clients/{id}": {
      "put": {
        "tags": [
          "Clients"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "company_name": {
                    "example": "any"
                  },
                  "company_type": {
                    "example": "any"
                  },
                  "company_size": {
                    "example": "any"
                  },
                  "company_industry": {
                    "example": "any"
                  },
                  "tax_info": {
                    "example": "any"
                  },
                  "address_line1": {
                    "example": "any"
                  },
                  "address_line2": {
                    "example": "any"
                  },
                  "country": {
                    "example": "any"
                  },
                  "state": {
                    "example": "any"
                  },
                  "city": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/clients/clients/{id}/deactivate": {
      "patch": {
        "tags": [
          "Clients"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}